requirements.txt
Django==4.2.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djoser==2.2.2
PyYAML==6.0.1
celery==5.3.6
redis==5.0.1